+++ tests/test_model.py
@@ -1,0 +1,37 @@
+# Macro-tests. Status: 0/13
+
+# Set of models
+# Set of surveys
+# All ab's
+# Arbitrary angles
+# All transforms
+# All optimizations
+# Source/Receiver scaling
+
+# 0. __init__ (Ensure main functions are importable from empymod)
+
+# 1. bipole
+
+# 1.1. Comparison to analytical fullspace solution
+
+# 1.2. Comparison to analytical halfspace solution
+
+# 1.3. Comparison to EMmod
+
+# 1.4. Comparison to DIPOLE1D
+
+# 1.5. Comparison to Green3D
+
+# 2. dipole (Ensure it is equivalent to bipole)
+
+# 3. gpr (Check it remains as on paper)
+
+# 4. wavenumber (Finish wavenumber properly; write checks)
+
+# 5. frequency (Ensure it is equivalent to dipole with signal=None)
+
+# 6. time (Ensure its same as dipole with signal!=None)
+
+# 7. fem
+
+# 8. tem