+++ tests/test_utils.py
@@ -1,0 +1,43 @@
+# utils. Status: 0/21
+
+# 1. EMArray
+
+# 2. check_time
+
+# 3. check_model
+
+# 4. check_frequency
+
+# 5. check_hankel
+
+# 6. check_opt
+
+# 7. check_dipole
+
+# 8. check_bipole
+
+# 9. check_ab
+
+# 10. get_abs
+
+# 11. get_geo_fact
+
+# 12. get_azm_dip
+
+# 13. get_off_ang
+
+# 14. get_layer_nr
+
+# 15. printstartfinish
+
+# 16. conv_warning
+
+# 17. _check_shape
+
+# 18. _check_var
+
+# 19. _strvar
+
+# 20. _prnt_min_max_val
+
+# 21. _check_min